requirements.txt
flask
requests
beautifulsoup4
openpyxl
lxml
python-dotenv